+++ shopify/utils/tsup.config.ts
@@ -1,13 +1,13 @@
 import path from "node:path";
 import fs from "node:fs/promises";
-import jsDelivrImportsPlugin from "@repo/js-delivr-imports-plugin";
+// import jsDelivrImportsPlugin from "@repo/js-delivr-imports-plugin";
 import { defineConfig } from "tsup";
 import { generateDtsBundle } from "dts-bundle-generator";
 
 const outDir = "dist";
 
 const commonConfig = {
-  minify: true,
+  minify: false,
   bundle: false,
   splitting: true,
   clean: true,
@@ -56,27 +56,27 @@ export default defineConfig([
       await fs.writeFile(path.resolve(outDir, "main", "index.d.ts"), dts);
     },
   },
-  {
-    ...commonConfig,
-    format: "esm",
-    entry: ["src/**/*.ts"],
-    outDir: `${outDir}/jsdelivr`,
-    dts: false,
-    async onSuccess() {
-      const [indexDts] = generateDtsBundle([
-        {
-          filePath: path.resolve("./src/index.ts"),
-        },
-      ]);
+  // {
+  //   ...commonConfig,
+  //   format: "esm",
+  //   entry: ["src/**/*.ts"],
+  //   outDir: `${outDir}/jsdelivr`,
+  //   dts: false,
+  //   async onSuccess() {
+  //     const [indexDts] = generateDtsBundle([
+  //       {
+  //         filePath: path.resolve("./src/index.ts"),
+  //       },
+  //     ]);
 
-      if (!indexDts) return;
+  //     if (!indexDts) return;
 
-      await fs.writeFile(
-        path.resolve(outDir, "jsdelivr", "index.d.ts"),
-        indexDts
-      );
-    },
+  //     await fs.writeFile(
+  //       path.resolve(outDir, "jsdelivr", "index.d.ts"),
+  //       indexDts
+  //     );
+  //   },
 
-    esbuildPlugins: [jsDelivrImportsPlugin()],
-  },
+  //   esbuildPlugins: [jsDelivrImportsPlugin()],
+  // },
 ]);