+++ shopify/packages/hybrid-cart-api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @brytdesigns/shopify-hybrid-cart-api
 
+## 1.1.2
+
+### Patch Changes
+
+- Removed jsdelivr custom build need
+- Updated dependencies
+  - @brytdesigns/shopify-utils@1.1.1
+
 ## 1.1.1
 
 ### Patch Changes