+++ shopify/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @brytdesigns/shopify-utils
 
+## 1.1.1
+
+### Patch Changes
+
+- Removed jsdelivr custom build need
+
 ## 1.1.0
 
 ### Minor Changes